+++ tests/Mutable/Enhanced/MapTest.php
@@ -51,7 +51,6 @@ class MapTest extends \MF\Collection\Mutable\MapTest
 
     /**
      * @param callable|string $reducer
-     * @param array $values
      * @param mixed $expected
      *
      * @dataProvider reduceByArrowFunctionProvider
@@ -85,7 +84,6 @@ class MapTest extends \MF\Collection\Mutable\MapTest
 
     /**
      * @param callable|string $reducer
-     * @param array $values
      * @param mixed $initialValue
      * @param mixed $expected
      *