+++ tests/Mutable/Generic/ListTest.php
@@ -64,7 +64,6 @@ class ListTest extends AbstractTestCase
 
     /**
      * @param string $valueType
-     * @param array $values
      *
      * @dataProvider validValuesProvider
      */