+++ tests/Mutable/ListTest.php
@@ -36,7 +36,6 @@ class ListTest extends AbstractTestCase
     }
 
     /**
-     * @param array $array
      * @param bool $recursive
      *
      * @dataProvider arrayProvider
@@ -158,8 +157,6 @@ class ListTest extends AbstractTestCase
     }
 
     /**
-     * @param array $array
-     *
      * @dataProvider arrayProvider
      */
     public function testShouldGetCount(array $array): void
@@ -394,8 +391,6 @@ class ListTest extends AbstractTestCase
     }
 
     /**
-     * @param callable $reducer
-     * @param array $values
      * @param mixed $expected
      *
      * @dataProvider reduceProvider
@@ -433,8 +428,6 @@ class ListTest extends AbstractTestCase
     }
 
     /**
-     * @param callable $reducer
-     * @param array $values
      * @param mixed $initialValue
      * @param mixed $expected
      *